pyproject.toml
[tool.poetry]
name = "python-web-hw11"
version = "0.1.0"
description = ""
authors = ["Andriy <AndriySydor91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {extras = ["email"], version = "^2.8.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
fastapi-limiter = "^0.1.6"
aioredis = "^2.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
